Made a change where NOAA API is only fetched if local db is empty
Implemented a change where the weather API is only fetched if the local database is currently unpopulated. Mechanism must be added later to update the data at the same interval that is done by the NOAA. Increase the amount of data point retrieved for the given range by reducing interval to 5 minutes instead of 15 minutes.